Add files via upload
ftpcon.py - This script will check for the anonymous login in the target's  machine. 
pxsshpy.py - This script is a simple demonstration of pxssh library. In this code the scripts takes target address,username and password and make ssh connection with the target.
scanandsearch.py - This script is a simple script which scan for only essential ports and check if they are open. 
ftpc.py - This script is ftp bruteforcing script. 
sshrbruter.py - This script is used to brute force username and password in SSH port.
sshpy.py - This script is the demonstration of usage of paramiko library in python.